+++ yash-env/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - External dependency versions:
     - Rust 1.70.0 → 1.77.0
+- Internal dependency versions:
+    - annotate-snippets 0.10.0 → 0.11.4
 
 ### Fixed